implementacja logiki biznesowej

+++ src/main/java/pl/facility_rental/facility/business/FacilityService.java
@@ -1,8 +1,10 @@
 package pl.facility_rental.facility.business;
 
 import org.springframework.stereotype.Service;
+import org.springframework.web.context.annotation.RequestScope;
 import pl.facility_rental.facility.data.FacilityRepository;
 import pl.facility_rental.facility.model.MongoSportsFacility;
+import pl.facility_rental.rent.business.RentService;
 import pl.facility_rental.rent.data.RentRepository;
 
 import java.util.List;
@@ -10,13 +12,15 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequestScope
 public class FacilityService {
 
     private final FacilityRepository facilityRepository;
+    private final RentService rentService;
 
-
-    public FacilityService(FacilityRepository facilityRepository) {
+    public FacilityService(FacilityRepository facilityRepository, RentService rentService) {
         this.facilityRepository = facilityRepository;
+        this.rentService = rentService;
     }
 
     public List<SportsFacility> findAll() {
@@ -36,8 +40,12 @@ public class FacilityService {
     }
 
     public SportsFacility deleteById(UUID id) throws Exception {
+        if(!rentService.findRentsForFacility(id).isEmpty()) {
+            throw new Exception("There are still booked rents for this facility!");
+        }
         return facilityRepository.delete(id);
     }
+
 
 
 }
